+++ actions/views/drafts.ts
@@ -1,6 +1,8 @@
 // Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
 // See LICENSE.txt for license information.
 
+import {batchActions} from 'redux-batched-actions';
+
 import {DispatchFunc, GetStateFunc} from 'mattermost-redux/types/actions';
 import {getCurrentUserId} from 'mattermost-redux/selectors/entities/users';
 import {syncedDraftsAreAllowedAndEnabled} from 'mattermost-redux/selectors/entities/preferences';
@@ -12,7 +14,29 @@ import type {GlobalState} from 'types/store';
 import {PostDraft} from 'types/store/draft';
 import {getGlobalItem} from 'selectors/storage';
 
+import type {Draft as ServerDraft} from '@mattermost/types/drafts';
 import type {UserProfile} from '@mattermost/types/users';
+import {StoragePrefixes} from 'utils/constants';
+import {FileInfo} from '@mattermost/types/files';
+
+type Draft = {
+    key: keyof GlobalState['storage']['storage'];
+    value: PostDraft;
+    timestamp: Date;
+}
+
+export function getDrafts(teamId: string) {
+    return async (dispatch: DispatchFunc) => {
+        let drafts: ServerDraft[] = [];
+        drafts = await Client4.getUserDrafts(teamId);
+        const actions = (drafts || []).map((draft) => {
+            const {key, value} = transformServerDraft(draft);
+            return setGlobalItem(key, value);
+        });
+
+        dispatch(batchActions(actions));
+    };
+}
 
 export function removeDraft(key: string, channelId: string, rootId = '') {
     return async (dispatch: DispatchFunc, getState: GetStateFunc) => {
@@ -69,4 +93,33 @@ function upsertDraft(draft: PostDraft, userId: UserProfile['id'], rootId = '', c
     };
 
     return Client4.upsertDraft(newDraft, connectionId);
+}
+
+export function transformServerDraft(draft: ServerDraft): Draft {
+    let key: Draft['key'] = `${StoragePrefixes.DRAFT}${draft.channel_id}`;
+
+    if (draft.root_id !== '') {
+        key = `${StoragePrefixes.COMMENT_DRAFT}${draft.root_id}`;
+    }
+
+    let fileInfos: FileInfo[] = [];
+    if (draft.metadata?.files) {
+        fileInfos = draft.metadata.files;
+    }
+
+    return {
+        key,
+        timestamp: new Date(draft.update_at),
+        value: {
+            message: draft.message,
+            fileInfos,
+            props: draft.props,
+            uploadsInProgress: [],
+            channelId: draft.channel_id,
+            rootId: draft.root_id,
+            createAt: draft.create_at,
+            updateAt: draft.update_at,
+            show: true,
+        },
+    };
 }
